Add RollData structure to have per-roll stuff
Having InternalContext on RollerConfig made it non-thread-safe, so if
multiple rolls used the same RollerConfig they would be stepping on each
other. Moving this to RollData allows us to be more safe. RollData also
allows per-roll functions/macros, so that we can e.g. expose the roll
macro only on rolls made as part of a RollPost even if the config is
used for non-RollPost rolls as well.

This is a breaking change, so bumping dev version to 3.0.0 instead.
Leaving ExecuteMacro as deprecated for now, to be removed in 4.0.0.

Began work on updating unit tests for code coverage, targeting 90% or
higher overall for the Dice and Dice.AST namespaces, and 80% for
Dice.Grammar. The new stuff still needs unit tests as well.

Adding this exposed a few bugs in some existing code, namely
FunctionRegistry.Remove, which was fixed.