Update dependency versions
DiceRoller was built against System.ValueTuple 4.3.1 but the nuspec only
required 4.3.0, which means that a new install would fail due to
installing an outdated version.

While we're fixing that, also fix the version in the nuspec as well as
update ANTLR to 4.7.1 (as it's a bugfix release, this is not a breaking
change or new feature). Bump our version to 3.0.2 to compensate for
these updated dependencies.

No code changes have occurred.